# src/swagger/books.yaml
openapi: 3.0.0
info:
  title: Book API
  version: 1.0.0
  description: API documentation for the Book API

servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://qa-simple-book-api.vercel.app/api
    description: Production server

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Books
    description: Book management endpoints

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Book:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the book
        title:
          type: string
          description: Title of the book
        description:
          type: string
          description: Description of the book
        pages:
          type: integer
          nullable: true
          description: Number of pages in the book
      required:
        - id
        - title
        - description
        - pages

paths:
  /login:
    post:
      summary: Login and get a JWT token
      tags: [Auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The admin username
                password:
                  type: string
                  description: The admin password
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful, returns a JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: The JWT token
        '401':
          description: Unauthorized, invalid credentials
        '500':
          description: Internal server error

  /books:
    get:
      summary: Retrieve a list of all books
      tags: [Books]
      operationId: getAllBooks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '500':
          description: Internal server error
    
    post:
      summary: Create a new book
      tags: [Books]
      operationId: createBook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                pages:
                  type: integer
              required:
                - title
                - description
                - pages
      responses:
        '201':
          description: Book created successfully
        '400':
          description: Invalid input
        '500':
          description: Internal server error
    
    delete:
      summary: Delete all books
      tags: [Books]
      operationId: deleteAllBooks
      security:
        - bearerAuth: []
      responses:
        '204':
          description: All books deleted successfully
        '500':
          description: Internal server error

  /books/{id}:
    put:
      summary: Update a book
      tags: [Books]
      operationId: updateBook
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The book ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Book updated successfully
        '400':
          description: Invalid input
        '404':
          description: Book not found
        '500':
          description: Internal server error
    get:
      summary: Retrieve a specific book by its id
      tags: [Books]
      operationId: retrieveBook
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The book ID
      responses:
        '200':
          description: Book found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
        '500':
          description: Internal server error
  
    delete:
      summary: Delete a book
      tags: [Books]
      operationId: deleteBook
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The book ID
      responses:
        '204':
          description: Book deleted successfully
        '404':
          description: Book not found
        '500':
          description: Internal server error

  /books-pages/paginate:
    get:
      summary: Paginate books
      tags: [Books]
      operationId: paginateBooks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: The page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: The number of books per page
      responses:
        '200':
          description: A paginated list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '500':
          description: Internal server error

  /books/filter/title:
    get:
      summary: Filter books by title
      tags: [Books]
      operationId: filterBooksByTitle
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: title
          schema:
            type: string
          description: The title of the book
      responses:
        '200':
          description: A list of books filtered by title
        '500':
          description: Internal server error

  /books/filter/description:
    get:
      summary: Filter books by description
      tags: [Books]
      operationId: filterBooksByDescription
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: description
          schema:
            type: string
          description: The description of the book
      responses:
        '200':
          description: A list of books filtered by description
        '500':
          description: Internal server error